requirements.txt
dropbox==6.3.0
fudge==1.1.0
github3.py==0.9.5
httplib2==0.9.2
httpretty==0.8.14
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==1.1.1
pycrypto==2.6.1
PyInstaller==3.2
PyTumblr==0.0.6
requests==2.10.0
requests-oauthlib==0.6.1
scrypt==0.7.1
simplejson==3.8.2
six==1.10.0
soundcloud==0.5.0
twython==3.4.0
uritemplate.py==0.3.0
urllib3==1.15.1
